Completed refactoring to remove cyclical imports and exports
