+++ Slime_game/classes/Entities/Enemies/Verm.js
@@ -2,7 +2,7 @@ import { StateMachine, RectangularTriggerRegion, Trigger } from "../../../librar
 import {aStar} from "../../Pathing.js";
 import { Enemy } from "../Enemy.js";
 import { PatrolState, FleeState, HideState} from "./VermStates.js";
-import { currentLevel } from "../../Controller.js";
+import { currentLevel } from "../../Global.js";
 
 //The Enemy is an object that will contain unique methods allowing player interaction
 class Verm extends Enemy {
